waymo_env_requirements.txt
absl-py==1.4.0
astunparse==1.6.3
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
cachetools==5.5.2
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
cloud-init==25.2
cloudpickle==3.1.1
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.3.2
cryptography==3.4.8
cycler==0.12.1
dacite==1.8.1
dask==2023.3.1
dataclass_array==1.5.1
dbus-python==1.2.18
decorator==5.2.1
distlib==0.4.0
distro==1.7.0
distro-info==1.1+ubuntu0.2
dm-tree==0.1.9
einops==0.8.1
einsum==0.3.0
etils==1.13.0
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.9.0
gast==0.4.0
google-auth==2.16.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.74.0
h5py==3.15.1
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
imageio==2.37.0
immutabledict==2.2.0
importlib-metadata==4.6.4
importlib_resources==6.5.2
incremental==21.3.0
intel_tensorflow==2.14.0
jax==0.4.13
jaxlib==0.4.13
jeepney==0.7.1
Jinja2==3.0.3
joblib==1.5.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keras==2.13.1
keyring==23.5.0
kiwisolver==1.4.9
lark==1.3.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
libclang==18.1.1
locket==1.0.0
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.6.1
mdurl==0.1.2
ml-dtypes==0.2.0
more-itertools==8.10.0
namex==0.1.0
netifaces==0.11.0
networkx==3.4.2
numpy==1.23.5
oauthlib==3.2.0
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==1.5.3
partd==1.4.2
Pillow==9.2.0
platformdirs==4.5.0
plotly==5.13.1
protobuf==4.25.8
pyarrow==16.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycurl==7.44.1
Pygments==2.19.2
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
pytz==2022.1
PyWavelets==1.8.0
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==2.0.0
rich==14.2.0
rsa==4.9.1
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.15.3
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
systemd-python==234
tenacity==9.1.2
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-probability==0.21.0
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2025.5.10
toolz==1.1.0
tqdm==4.67.1
Twisted==22.1.0
typing_extensions==4.5.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
virtualenv==20.35.3
visu3d==1.5.1
wadllib==1.3.6
waymo-open-dataset-tf-2-12-0==1.6.7
Werkzeug==3.1.3
wrapt==1.14.2
zipp==1.0.0
zope.interface==5.4.0
